+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+## [0.3.1] - 2024-02-06
+
+- Improvement in steroid recipe new_project : Code improvement (variable name for boolean_choices)
+- Improvement in steroid generator : Add `require 'tty/prompt'` by default
+
 ## [0.3.0] - 2024-02-05
 
 - Improvement in steroid recipe new_project : Use of TTY-prompt for better interaction